# lib/api-zod/src/generated/types/story.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StoryCategory } from "./storyCategory";

export interface Story {
  id: string;
  title: string;
  category: StoryCategory;
  description: string;
  content: string;
  coverEmoji: string;
  videoUrl?: string | null;
  videoFile?: string | null;
  quizEnabled: boolean;
  audioFile?: string | null;
  images: string[];
  voiceRecordings: string[];
  isFavorite: boolean;
  readCount: number;
  createdAt: number;
}